Update dependency key for urdfdom-headers.
We're shadowing this key for r2b2 since upstream is out of date but we should
use the rosdep key so we can eventually target upstream.

https://github.com/ros2/robot_model/pulls/1

This is the second attempt to do so after the first failed due to the
urdfdom_headers package not being renamed for source builds. That is
being resolved by placing a package.xml with the updated name in the
source tree rather than in a release patch.